gen:rename packages
