--- modules/gen/src/main/scala/gem/sql/MagnitudeBand.scala
@@ -1,12 +1,0 @@
-// Copyright (c) 2016-2019 Association of Universities for Research in Astronomy, Inc. (AURA)
-// For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
-
-package gem.sql
-
-/** Minimal MagnitudeBand classes to support reading from the enum
-  * tables.
-  */
-
-final case class MagnitudeBand(id: String)
-
-final case class MagnitudeSystem(id: String)
