+++ modules/gen/src/main/scala/gsp/core/gen/GnirsPixelScale.scala
@@ -1,12 +1,10 @@
 // Copyright (c) 2016-2019 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-package gem
+package gsp.core.gen
 
-import doobie._
-import java.time.ZoneId
+/** Minimal PixelScale classes to support reading from the enum
+  * tables.
+  */
 
-package object sql {
-  implicit val ZoneIdMeta: Meta[ZoneId] =
-    Meta[String].timap(ZoneId.of)(_.toString)
-}
+final case class GnirsPixelScale(id: String)
