+++ modules/gen/src/main/scala/gsp/core/gen/KeywordName.scala
@@ -1,6 +1,6 @@
 // Copyright (c) 2016-2019 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-package gem.sql
+package gsp.core.gen
 
 final case class KeywordName(id: String)
