+++ modules/gen/src/main/scala/gsp/core/gen/MagnitudeBand.scala
@@ -1,10 +1,12 @@
 // Copyright (c) 2016-2019 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-package gem.sql
+package gsp.core.gen
 
-/** Minimal PixelScale classes to support reading from the enum
+/** Minimal MagnitudeBand classes to support reading from the enum
   * tables.
   */
 
-final case class GnirsPixelScale(id: String)
+final case class MagnitudeBand(id: String)
+
+final case class MagnitudeSystem(id: String)
