+++ modules/gen/src/main/scala/gsp/core/gen/MagnitudeValue.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2016-2019 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-package gem.sql
+package gsp.core.gen
 
 import java.math.RoundingMode.HALF_UP
