+++ modules/gen/src/main/scala/gsp/core/gen/Main.scala
@@ -1,10 +1,10 @@
 // Copyright (c) 2016-2019 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-package gem.sql
+package gsp.core.gen
 
 import doobie._, doobie.implicits._
-import gem.sql.enum._
+import gsp.core.gen.enum._
 import java.nio.file._
 import cats.implicits._, cats.effect.{ IO, ContextShift }
