+++ modules/gen/src/main/scala/gsp/core/gen/enum/GsaoiEnums.scala
@@ -1,13 +1,13 @@
 // Copyright (c) 2016-2019 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-package gem.sql
+package gsp.core.gen
 package enum
 
 import doobie._, doobie.implicits._
 import shapeless.record._
 import shapeless.Witness
-import gem.sql.FiniteDuration
+import gsp.core.gen.FiniteDuration
 
 object GsaoiEnums {
   import EnumRefs._
