+++ modules/gen/src/main/scala/gsp/core/gen/enum/MiscEnums.scala
@@ -1,14 +1,14 @@
 // Copyright (c) 2016-2019 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-package gem.sql
+package gsp.core.gen
 package enum
 
 import doobie._
 import doobie.implicits._
 import java.time.ZoneId
 
-import gem.sql.EnumDef
+import gsp.core.gen.EnumDef
 import shapeless.record._
 import shapeless.Witness
