+++ modules/gen/src/main/scala/gsp/core/gen/enum/TargetEnums.scala
@@ -1,7 +1,7 @@
 // Copyright (c) 2016-2019 Association of Universities for Research in Astronomy, Inc. (AURA)
 // For license information see LICENSE or https://opensource.org/licenses/BSD-3-Clause
 
-package gem.sql
+package gsp.core.gen
 package enum
 
 import doobie._, doobie.implicits._
